moved IConfigurer to Sienar.Configuration namespace
