+++ src/Sienar.Architecture.Web/Plugins/MvcPlugin.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Sienar.Configuration;
 using Sienar.Infrastructure;
 
 namespace Sienar.Plugins;
