+++ src/Sienar.Plugin.Cms.Server/Infrastructure/DefaultAuthenticationBuilderConfigurer.cs
@@ -3,6 +3,7 @@
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.DependencyInjection;
+using Sienar.Configuration;
 
 namespace Sienar.Infrastructure;
