+++ src/Sienar.Plugin.Cms.Server/Infrastructure/DefaultAuthenticationConfigurer.cs
@@ -2,6 +2,7 @@
 
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
+using Sienar.Configuration;
 
 namespace Sienar.Infrastructure;
