+++ src/Sienar.Plugin.Cms.Server/Infrastructure/DefaultAuthorizationConfigurer.cs
@@ -1,6 +1,7 @@
 ﻿#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
 using Microsoft.AspNetCore.Authorization;
+using Sienar.Configuration;
 
 namespace Sienar.Infrastructure;
