+++ src/Sienar.Ui/Plugins/MudBlazorPlugin.cs
@@ -1,4 +1,5 @@
 ﻿using MudBlazor.Services;
+using Sienar.Configuration;
 using Sienar.Infrastructure;
 
 namespace Sienar.Plugins;
