+++ src/Sienar.Utils/Configuration/IConfigurer.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using Microsoft.Extensions.Configuration;
 
-namespace Sienar.Infrastructure;
+namespace Sienar.Configuration;
 
 /// <summary>
 /// Used by Sienar to configure a service or middleware that is generally configured by an <see cref="Action{T}"/>
@@ -11,8 +10,8 @@ namespace Sienar.Infrastructure;
 public interface IConfigurer<TOptions> where TOptions : class
 {
 	/// <summary>
-	/// Configures an instance of <c>T</c>
+	/// Configures an instance of <c>TOptions</c>
 	/// </summary>
-	/// <param name="options">the instance of <c>T</c> to configure</param>
+	/// <param name="options">the instance of <c>TOptions</c> to configure</param>
 	void Configure(TOptions options);
 }
