+++ src/Sienar.Utils/Extensions/SienarUtilsServiceCollectionExtensions.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
+using Sienar.Configuration;
 using Sienar.Email;
 using Sienar.Hooks;
 using Sienar.Infrastructure;
